Work on settings

+++ src/Logic/Config/SeAdjustDisplayDurations.cs
@@ -4,7 +4,7 @@ namespace SubtitleAlchemist.Logic.Config;
 
 public class SeAdjustDisplayDurations
 {
-    public decimal AdjustDurationSeconds { get; set; } = 0.1m;
+    public double AdjustDurationSeconds { get; set; } = 0.1;
     public decimal AdjustDurationFixed { get; set; } = 0.1m;
     public int AdjustDurationPercent { get; set; } = 120;
     public string AdjustDurationLast { get; set; } = AdjustDurationModel.ModeSeconds;
